# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/value-set/page": {
      "get": {
        "operationId": "ValueSetController_pageValueSet",
        "parameters": [
          {
            "name": "parentSetCode",
            "required": false,
            "in": "query",
            "description": "父编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentSetName",
            "required": false,
            "in": "query",
            "description": "父名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortNumber",
            "required": false,
            "in": "query",
            "description": "排序号",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdUserId",
            "required": false,
            "in": "query",
            "description": "创建人id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdUserName",
            "required": false,
            "in": "query",
            "description": "创建人名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedUserId",
            "required": false,
            "in": "query",
            "description": "修改人id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "updatedUserName",
            "required": false,
            "in": "query",
            "description": "修改人姓名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "setCode",
            "required": false,
            "in": "query",
            "description": "集CODE",
            "schema": {
              "example": "SYS_GENDER",
              "type": "string"
            }
          },
          {
            "name": "setName",
            "required": false,
            "in": "query",
            "description": "集NAME",
            "schema": {
              "example": "性别",
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "值CODE",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "值NAME",
            "schema": {
              "example": "男",
              "type": "string"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ValueSetPageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询值集列表",
        "tags": ["值集模块"]
      }
    },
    "/value-set/schema-generator-holder-page-value-set": {
      "post": {
        "deprecated": true,
        "operationId": "ValueSetController__pageValueSet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetPageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["值集模块"]
      }
    },
    "/value-set/sets/page": {
      "get": {
        "operationId": "ValueSetController_pageValueSetGroups",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "setCode",
            "required": false,
            "in": "query",
            "description": "集CODE",
            "schema": {
              "example": "SYS_GENDER",
              "type": "string"
            }
          },
          {
            "name": "setName",
            "required": false,
            "in": "query",
            "description": "集NAME",
            "schema": {
              "example": "性别",
              "type": "string"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ValueSetGroupPageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "集维度分页查询（按 setCode 去重）",
        "tags": ["值集模块"]
      }
    },
    "/value-set/schema-generator-holder-sets-page-value-set": {
      "post": {
        "deprecated": true,
        "operationId": "ValueSetController__pageValueSetGroups",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetGroupPageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["值集模块"]
      }
    },
    "/value-set/option": {
      "get": {
        "operationId": "ValueSetController_pageOptionValueSet",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "返回的列表字段",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ValueSetPageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询值集列表(可选字段)",
        "tags": ["值集模块"]
      }
    },
    "/value-set/schema-generator-holder-page-option-value-set": {
      "post": {
        "deprecated": true,
        "operationId": "ValueSetController__pageOptionValueSet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetPageOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["值集模块"]
      }
    },
    "/value-set/by-set-codes": {
      "get": {
        "operationId": "ValueSetController_findValueSetBySetCodes",
        "parameters": [
          {
            "name": "setCodes",
            "required": true,
            "in": "query",
            "description": ",分隔的集CODE",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValueSetListRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按集code查询值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/schema-generator-holder-by-set-codes": {
      "post": {
        "deprecated": true,
        "operationId": "ValueSetController__findValueSetBySetCodes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["值集模块"]
      }
    },
    "/value-set/find/{id}": {
      "get": {
        "operationId": "ValueSetController_findValueSetById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValueSetRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/batch/query": {
      "get": {
        "operationId": "ValueSetController_findValueSetListByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "用“,”拼接的id",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValueSetListRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按ids查询值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/create": {
      "post": {
        "operationId": "ValueSetController_createValueSet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValueSetRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "创建值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/update/{id}": {
      "post": {
        "operationId": "ValueSetController_updateValueSet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueSetUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ValueSetRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/delete/{id}": {
      "post": {
        "operationId": "ValueSetController_removeValueSet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "删除值集",
        "tags": ["值集模块"]
      }
    },
    "/value-set/batch/delete": {
      "post": {
        "operationId": "ValueSetController_batchRemoveValueSet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量删除值集",
        "tags": ["值集模块"]
      }
    },
...
  },
  "info": {
    "title": "BACKED API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "head": {
            "$ref": "#/components/schemas/ResponseHeadDto"
          }
        },
        "required": ["head"]
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {}
      },
...
      "ValueSetPageRespDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "字典id"
          },
          "setCode": {
            "type": "string",
            "description": "集编码"
          },
          "setName": {
            "type": "string",
            "description": "集名称"
          },
          "code": {
            "type": "string",
            "description": "字典编码"
          },
          "name": {
            "type": "string",
            "description": "字典名称"
          },
          "parentSetCode": {
            "type": "string",
            "description": "父编码"
          },
          "parentSetName": {
            "type": "string",
            "description": "父名称"
          },
          "status": {
            "type": "number",
            "description": "状态：0-禁用，1-启用"
          },
          "sortNumber": {
            "type": "number",
            "description": "排序号"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间"
          },
          "createdUserId": {
            "type": "number",
            "description": "创建人id"
          },
          "createdUserName": {
            "type": "string",
            "description": "创建人名称"
          },
          "updatedUserId": {
            "type": "number",
            "description": "修改人id"
          },
          "updatedUserName": {
            "type": "string",
            "description": "修改人姓名"
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间"
          }
        },
        "required": ["id", "setCode", "setName", "code", "name"]
      },
...
      "ValueSetPageOptionDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数"
          },
          "keyword": {
            "type": "string",
            "description": "关键字",
            "example": ""
          },
          "fields": {
            "type": "string",
            "description": "返回的列表字段",
            "example": ""
          }
        }
      },
...
    }
  }
}